ultimos ajustes

+++ app/src/main/java/com/example/chargetech/adapters/ConsumoEnergeticoAdapter.kt
@@ -1,5 +1,8 @@
 package com.example.chargetech.adapters
 
+import android.app.AlertDialog
+import android.content.Context
+import android.content.Intent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -8,6 +11,7 @@ import android.widget.TextView
 import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.example.chargetech.R
+import com.example.chargetech.activities.ProfileActivity
 import com.example.chargetech.models.ConsumoEnergetico
 import com.example.chargetech.repositories.ConsumoEnergeticoRepository
 
@@ -39,19 +43,38 @@ class ConsumoEnergeticoAdapter(
             consumo.text = "${consumoEnergetico.consumo} kWh"
 
             deleteButton.setOnClickListener {
-                consumoEnergeticoRepository.deleteById(consumoEnergetico.id_consumo_energetico) { success ->
-                    if (success) {
-                        val position = adapterPosition
-                        if (position != RecyclerView.NO_POSITION) {
-                            consumos.removeAt(position)
-                            notifyItemRemoved(position)
-                        }
-                        Toast.makeText(itemView.context, "Consumo excluído com sucesso", Toast.LENGTH_SHORT).show()
-                    } else {
-                        Toast.makeText(itemView.context, "Erro ao excluir consumo", Toast.LENGTH_SHORT).show()
-                    }
-                }
+                showDeleteConfirmationDialog(consumoEnergetico)
             }
+        }
+
+        private fun showDeleteConfirmationDialog(consumoEnergetico: ConsumoEnergetico) {
+            val context = itemView.context
+            val alertDialog = AlertDialog.Builder(context)
+                .setTitle("Excluir consumo energético")
+                .setMessage("Você tem certeza que deseja excluir este consumo?")
+                .setPositiveButton("Sim") { dialog, _ ->
+                    consumoEnergeticoRepository.deleteById(consumoEnergetico.id_consumo_energetico) { success ->
+                        if (success) {
+                            val position = adapterPosition
+                            if (position != RecyclerView.NO_POSITION) {
+                                consumos.removeAt(position)
+                                notifyItemRemoved(position)
+                            }
+                            val profileIntent = Intent(context, ProfileActivity::class.java)
+                            context.startActivity(profileIntent)
+                            Toast.makeText(context, "Consumo excluído com sucesso", Toast.LENGTH_SHORT).show()
+                        } else {
+                            Toast.makeText(context, "Erro ao excluir consumo", Toast.LENGTH_SHORT).show()
+                        }
+                    }
+                    dialog.dismiss()
+                }
+                .setNegativeButton("Não") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                .create()
+
+            alertDialog.show()
         }
     }
 }
